--- a/benchmarks/cpp/utils/prepare_synthetic_data.py
+++ b/benchmarks/cpp/utils/prepare_synthetic_data.py
@@ -71,10 +71,13 @@ def token_norm_dist(root_args, **kwargs):
                 "max_output_len": max_output_len
             }, root_args.output)
     else:
-        print_text_dataset(
-            input_ids,
-            output_lens,
-        )
+        print_text_dataset(input_ids,
+                           output_lens,
+                           task_ids=task_ids if root_args.task_id != -1
+                           or root_args.rand_task_id is not None else None,
+                           lora_config={"lora_dir": root_args.lora_dir}
+                           if root_args.task_id != -1
+                           or root_args.rand_task_id is not None else None)
 
 
 @click.command()
@@ -141,7 +144,10 @@ def token_unif_dist(root_args, **kwargs):
                 "max_output_len": max_output_len
             }, root_args.output)
     else:
-        print_text_dataset(
-            input_ids,
-            output_lens,
-        )
+        print_text_dataset(input_ids,
+                           output_lens,
+                           task_ids=task_ids if root_args.task_id != -1
+                           or root_args.rand_task_id is not None else None,
+                           lora_config={"lora_dir": root_args.lora_dir}
+                           if root_args.task_id != -1
+                           or root_args.rand_task_id is not None else None)