--- a/benchmarks/cpp/utils/prepare_synthetic_data.py
+++ b/benchmarks/cpp/utils/prepare_synthetic_data.py
@@ -6,6 +6,22 @@ from utils.utils import (gen_random_tokens, get_norm_dist_lengths,
                          text_dataset_dump)
 
 
+def _generate_task_ids_and_lora_config(root_args, num_reqs):
+    """Generate task IDs and determine LoRA configuration based on root_args."""
+    if root_args.rand_task_id is None:
+        task_ids = [root_args.task_id for _ in range(num_reqs)]
+    else:
+        min_id, max_id = root_args.rand_task_id
+        task_ids = [random.randint(min_id, max_id) for _ in range(num_reqs)]
+
+    # Determine if task IDs and LoRA config should be used
+    use_task_config = root_args.task_id != -1 or root_args.rand_task_id is not None
+
+    return (task_ids, task_ids if use_task_config else None, {
+        "lora_dir": root_args.lora_dir
+    } if use_task_config else None)
+
+
 @click.command()
 @click.option("--num-requests",
               required=True,
@@ -33,7 +49,6 @@ def token_norm_dist(root_args, **kwargs):
     input_ids = []
     input_lens = []
     output_lens = []
-    task_ids = []
 
     input_lens = get_norm_dist_lengths(kwargs['input_mean'],
                                        kwargs['input_stdev'],
@@ -51,11 +66,8 @@ def token_norm_dist(root_args, **kwargs):
     input_ids = gen_random_tokens(input_lens, root_args.tokenizer,
                                   root_args.random_seed)
 
-    if root_args.rand_task_id is None:
-        task_ids = [root_args.task_id for _ in range(num_reqs)]
-    else:
-        min_id, max_id = root_args.rand_task_id
-        task_ids = [random.randint(min_id, max_id) for _ in range(num_reqs)]
+    task_ids, print_task_ids, lora_config = _generate_task_ids_and_lora_config(
+        root_args, num_reqs)
 
     if not root_args.std_out:
         text_dataset_dump(
@@ -73,11 +85,8 @@ def token_norm_dist(root_args, **kwargs):
     else:
         print_text_dataset(input_ids,
                            output_lens,
-                           task_ids=task_ids if root_args.task_id != -1
-                           or root_args.rand_task_id is not None else None,
-                           lora_config={"lora_dir": root_args.lora_dir}
-                           if root_args.task_id != -1
-                           or root_args.rand_task_id is not None else None)
+                           task_ids=print_task_ids,
+                           lora_config=lora_config)
 
 
 @click.command()
@@ -107,7 +116,6 @@ def token_unif_dist(root_args, **kwargs):
     input_ids = []
     input_lens = []
     output_lens = []
-    task_ids = []
 
     input_lens = get_unif_dist_lengths(kwargs['input_min'], kwargs['input_max'],
                                        kwargs['num_requests'],
@@ -124,11 +132,8 @@ def token_unif_dist(root_args, **kwargs):
     input_ids = gen_random_tokens(input_lens, root_args.tokenizer,
                                   root_args.random_seed)
 
-    if root_args.rand_task_id is None:
-        task_ids = [root_args.task_id for _ in range(num_reqs)]
-    else:
-        min_id, max_id = root_args.rand_task_id
-        task_ids = [random.randint(min_id, max_id) for _ in range(num_reqs)]
+    task_ids, print_task_ids, lora_config = _generate_task_ids_and_lora_config(
+        root_args, num_reqs)
 
     if not root_args.std_out:
         text_dataset_dump(
@@ -146,8 +151,5 @@ def token_unif_dist(root_args, **kwargs):
     else:
         print_text_dataset(input_ids,
                            output_lens,
-                           task_ids=task_ids if root_args.task_id != -1
-                           or root_args.rand_task_id is not None else None,
-                           lora_config={"lora_dir": root_args.lora_dir}
-                           if root_args.task_id != -1
-                           or root_args.rand_task_id is not None else None)
+                           task_ids=print_task_ids,
+                           lora_config=lora_config)