--- a/tensorrt_llm/llmapi/llm_args.py
+++ b/tensorrt_llm/llmapi/llm_args.py
@@ -1259,7 +1259,7 @@ class BaseLlmArgs(BaseModel):
                     "lora_dir is empty, so custom embedding or lm head will not be applied."
                 )
 
-        if self.enable_lora and self.lora_config is not None and self.backend == 'pytorch':
+        if not self.enable_lora and self.lora_config is not None and self.backend == 'pytorch':
             logger.warning(
                 "enable_lora is ignored when lora_config is provided for pytorch backend."
             )
@@ -1763,6 +1763,7 @@ def update_llm_args_with_extra_dict(
         "extended_runtime_perf_knob_config": ExtendedRuntimePerfKnobConfig,
         "pytorch_backend_config": PyTorchConfig,
         "cache_transceiver_config": CacheTransceiverConfig,
+        "lora_config": LoraConfig,
     }
     for field, field_type in field_mapping.items():
         if field in llm_args_dict: