# Summary
This pull request adds end-to-end LoRA support for PyTorch-based performance tests in the TRTLLM benchmark suite. It introduces new configuration parameters, command-line handling, and test cases so that LoRA checkpoints can be symlinked, passed into data-generation scripts, and automatically included in the YAML config for model instantiation.

This is 

- [ ] Enable bench-side and other changes needed for test harness to enable running lora with pyt (#4848 )
- [x] Test harness plumbing to automate the prepare_data calls with lora requests in it + `extra-llm-api-options.yml` to include appropriate lora metadata
- [x] Add and Validate a basic e2e perftest using `llama-3.1-8b-fp8` with `chinese-llama-2-lora-13b` (numbers below)


---

## Perf summary of `llama_v3.1_8b_instruct` FP8 (PyTorch 1-LoRA on/off) — 1 × H100

### Invariants  
| **Invariant** | Value |
|---------------|-------|
| GPU(s) / backend / benchmarking runtime | 1 / PyTorch / trtllm-bench |
| Model checkpoint | *llama_v3.1-8b-instruct-FP8* |
| ISL / OSL  | **128 × 128** |
| Requests | 100 |
| Concurrency | 2 |


### Throughput metrics

| LoRA | Req/s | Out TPS<br>(tok/s) | Tok TPS<br>(tok/s) | Total Latency<br>(ms) | Avg Latency<br>(ms) | TPS / User | TPS / GPU |
|:---:|------:|-------------------:|-------------------:|----------------------:|--------------------:|-----------:|----------:|
| 0 | **3.5831** | **458.63** | **917.27** | 27 909 | **558.05** | **229.37** | **458.63** |
| 1 | 2.1781 | 264.77 | 539.85 | 45 911 | 916.75 | 132.96 | 264.77 |

---

### Latency percentiles (ms)

| LoRA | P50 | P90 | P95 | P99 | Min | Max |
|:---:|----:|----:|----:|----:|----:|----:|
| 0 | **557.36** | 561.02 | 562.65 | 569.14 | 554.78 | 569.14 |
| 1 | 909.45 | 927.66 | 1 113.07 | 1 135.33 | 717.88 | 1 135.33 |


### LoRA Impact
_For the above perf config settings, adding one LoRA adapter slashes output TPS by ≈ 42 % (458 → 265 tok/s) while driving average latency up ≈ 64 % (0.56 s → 0.92 s)._